Update: Separation of code into modules
