+++ src/componentTree.ts
@@ -1,23 +1,7 @@
 import * as fs from 'fs';
-import * as fspath from 'path';
-import ts from 'typescript';
 import * as vscode from 'vscode';
-import { findAbsoluteFilePath } from './utils';
-
-const STENCIL_CONFIG_FILE_NAME = 'stencil.config.ts';
-const DOCS_JSON_FILE_NAME = 'docs.json';
-
-type Component = {
-    tag: string;
-    docs: string;
-    filePath: string;
-    dependents: string[];
-    dependencies: string[];
-};
-
-type DocsJson = {
-    components: Component[];
-}
+import { findAbsoluteFilePath } from './fileUtils';
+import { DocsJson } from './types';
 
 export class ComponentTreeDataProvider implements vscode.TreeDataProvider<Node> {
     private rootNodes: Node[] = [];
@@ -25,16 +9,16 @@ export class ComponentTreeDataProvider implements vscode.TreeDataProvider<Node> 
     private _onDidChangeTreeData: vscode.EventEmitter<Node | undefined | void> = new vscode.EventEmitter<Node | undefined | void>();
     readonly onDidChangeTreeData: vscode.Event<Node | undefined | void> = this._onDidChangeTreeData.event;
 
-    constructor(private path: string | undefined) {
+    constructor(private docsJsonPath: string | undefined) {
     }
 
-    refresh(path: string | undefined): void {
-        this.path = path;
+    refresh(docsJsonPath: string): void {
+        this.docsJsonPath = docsJsonPath;
         this._onDidChangeTreeData.fire();
     }
 
-    getPath(): string | undefined {
-        return this.path;
+    getDocsJsonPath(): string | undefined {
+        return this.docsJsonPath;
     }
 
     getTreeItem(element: Node): vscode.TreeItem | Thenable<vscode.TreeItem> {
@@ -42,17 +26,14 @@ export class ComponentTreeDataProvider implements vscode.TreeDataProvider<Node> 
     }
 
     getChildren(element?: Node): vscode.ProviderResult<Node[]> {
-        if (!this.path) {
-            console.log('[ComponentTreeDataProvider] No defined path.');
-            this.setDocsJsonNotFound();
-            return;
-        }
-
-        const docsJson = this.getDocsJson(this.path);
-
         if (element) {
             return this.getElementNodesFromRootNodes(this.rootNodes, element);
         } else {
+            if (!this.docsJsonPath) {
+                console.log('[ComponentTreeDataProvider] No docsJsonPath defined.');
+                return;
+            }
+            const docsJson = JSON.parse(fs.readFileSync(this.docsJsonPath, 'utf8'));
             this.rootNodes = this.getRootNodesFromDocsJson(docsJson);
             return this.rootNodes;
         }
@@ -85,121 +66,13 @@ export class ComponentTreeDataProvider implements vscode.TreeDataProvider<Node> 
 
         return components.map(component => {
             const { tag, docs, filePath, dependents, dependencies } = component;
-            const path = this.getAbsoluteFilePath(filePath);
+            const path = findAbsoluteFilePath(this.docsJsonPath, filePath);
             const filteredDependencies = dependencies.filter(dependency => components.find(component => component.tag === dependency));
             const collapsibleState = filteredDependencies.length > 0
                 ? vscode.TreeItemCollapsibleState.Collapsed : vscode.TreeItemCollapsibleState.None;
 
             return new Node(tag, docs, path, dependents, filteredDependencies, collapsibleState);
         });
-    }
-
-    private getAbsoluteFilePath(relativeFilePath: string): string | undefined {
-        if (this.path && !this.path.endsWith(DOCS_JSON_FILE_NAME)) {
-            return fspath.join(this.path, relativeFilePath);
-        }
-
-        return findAbsoluteFilePath(this.path, relativeFilePath);
-    }
-
-    private getDocsJson(path: string): DocsJson | undefined {
-        try {
-            if (!path.endsWith(DOCS_JSON_FILE_NAME)) {
-                const docsJsonPath = fspath.join(path, DOCS_JSON_FILE_NAME);
-                if (fs.existsSync(docsJsonPath)) {
-                    path = docsJsonPath;
-                }
-                else {
-                    const stencilConfigPath = fspath.join(path, STENCIL_CONFIG_FILE_NAME);
-                    const relativeDocsJsonPath = this.getDocsJsonPath(stencilConfigPath);
-                    if (!relativeDocsJsonPath) {
-                        throw Error('Path not found.');
-                    }
-                    path = fspath.join(path, relativeDocsJsonPath);
-                }
-            }
-            if (fs.existsSync(path)) {
-                return JSON.parse(fs.readFileSync(path, 'utf-8'));
-            } else {
-                throw Error('File not found.');
-            }
-        } catch (error) {
-            console.error(`[${DOCS_JSON_FILE_NAME}] ${error}`);
-            this.setDocsJsonNotFound();
-        }
-    }
-
-    private getDocsJsonPath(stencilConfigPath: string): string | undefined {
-        try {
-            if (fs.existsSync(stencilConfigPath)) {
-                const stencilConfigFile = fs.readFileSync(stencilConfigPath, 'utf-8');
-                const stencilConfigSourceFile = ts.createSourceFile(STENCIL_CONFIG_FILE_NAME, stencilConfigFile, ts.ScriptTarget.ESNext, false);
-
-                return this.findDocsJsonPath(stencilConfigSourceFile);
-            } else {
-                throw Error('File not found.');
-            }
-        } catch (error) {
-            console.error(`[${STENCIL_CONFIG_FILE_NAME}] ${error}`);
-        }
-    }
-
-    private findDocsJsonPath(stencilConfigSourceFile: ts.SourceFile): string {
-        let docsJsonPath: string = '';
-
-        stencilConfigSourceFile.forEachChild(node => {
-            if (
-                ts.isVariableStatement(node) &&
-                node.declarationList.declarations.length > 0
-            ) {
-                const declaration = node.declarationList.declarations[0];
-
-                if (
-                    ts.isIdentifier(declaration.name) &&
-                    declaration.name.text === 'config' &&
-                    declaration.initializer &&
-                    ts.isObjectLiteralExpression(declaration.initializer)
-                ) {
-                    const outputTargetsProperty = declaration.initializer.properties.find(prop =>
-                        ts.isPropertyAssignment(prop) &&
-                        ts.isIdentifier(prop.name) &&
-                        prop.name.text === 'outputTargets'
-                    ) as ts.PropertyAssignment | undefined;
-
-                    if (outputTargetsProperty && ts.isArrayLiteralExpression(outputTargetsProperty.initializer)) {
-                        const outputTargetsArray = outputTargetsProperty.initializer.elements;
-
-                        for (const element of outputTargetsArray) {
-                            if (ts.isObjectLiteralExpression(element)) {
-                                const typeProperty = element.properties.find(prop =>
-                                    ts.isPropertyAssignment(prop) &&
-                                    ts.isIdentifier(prop.name) &&
-                                    prop.name.text === 'type' &&
-                                    ts.isStringLiteral(prop.initializer) &&
-                                    prop.initializer.text === 'docs-json'
-                                );
-
-                                if (typeProperty) {
-                                    element.properties.forEach(prop => {
-                                        if (ts.isPropertyAssignment(prop) && ts.isIdentifier(prop.name) && prop.name.text === 'file') {
-                                            docsJsonPath = ts.isStringLiteral(prop.initializer)
-                                                ? prop.initializer.text
-                                                : prop.initializer.getText();
-                                        }
-                                    });
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        });
-
-        return docsJsonPath;
-    }
-
-    private setDocsJsonNotFound() {
-        vscode.commands.executeCommand('setContext', 'extension.docsJsonNotFound', true);
     }
 }
