+++ src/constants.ts
@@ -1,0 +1,5 @@
+export const JSON_EXTENSION_NAME = '.json';
+
+export const DOCS_JSON_FILE_NAME = `docs.json`;
+
+export const STENCIL_CONFIG_FILE_NAME = 'stencil.config.ts';
