+++ src/extension.ts
@@ -4,15 +4,18 @@
  * @description application will start from this file
  */
 import * as vscode from 'vscode';
-import { ComponentTreeDataProvider } from './componentTree';
 import { comands } from './comands';
+import { ComponentTreeDataProvider } from './componentTree';
+import { getDocsJsonPath } from './fileUtils';
 
 export function activate(context: vscode.ExtensionContext) {
 	const rootPath = vscode.workspace.workspaceFolders && (vscode.workspace.workspaceFolders.length > 0)
 		? vscode.workspace.workspaceFolders[0].uri.fsPath
 		: undefined;
 
-	const componentTreeDataProvider = new ComponentTreeDataProvider(rootPath);
+	const docsJsonPath = getDocsJsonPath(rootPath);
+
+	const componentTreeDataProvider = new ComponentTreeDataProvider(docsJsonPath);
 	vscode.window.registerTreeDataProvider('treeView', componentTreeDataProvider);
 
 	comands.forEach(command => {
