+++ src/types.ts
@@ -1,0 +1,11 @@
+type Component = {
+    tag: string;
+    docs: string;
+    filePath: string;
+    dependents: string[];
+    dependencies: string[];
+};
+
+export type DocsJson = {
+    components: Component[];
+}
